To get the first build to work I had to change the project build settings to my name or apple id. You may need to do that too.
ML part
This uses an on-phone ML model.
It crops the center of the screen and queries the ML model.
It displays on a temporary Blue label(we'll remove this UI label later) the recognized object name and confidence.
Problem: the model is often wrong, but that's funny for now
Problem: sometimes xcode doesn't recognize that the inception model is there, and I need to retype it then it's fine.

AR part
If you tap on an object in the live-camera feed, it creates an augmented reality text string label using the latest recognized object name from the ML model. It tries to place that AR label on the closest object in the screen.
Problem: the AR label is created and put into the view but it's placed above and to the right of the object.
Problem: the AR labels don't rotate to face you. So some text ends up backwards.

I had to recreate the camera roll and camera capture buttons. I am not very good at that, so I don't think the constraints I put on the buttons are the best. I hacked around until they were in the right spot.
It took me a while to learn how to have the AR view instead of the UI Image or camera view and have buttons and labels on top. It ended up being that I had to make a View, then add the AR View, buttons, and labels on top of that original View. You'll see that hierarchy in the storyboard.